% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdateEthio-package.R
\docType{package}
\name{cdateEthio-package}
\alias{cdateEthio}
\alias{cdateEthio-package}
\title{cdateEthio: Convert date from Ethiopian calendar to Gregorian.}
\description{
The function convert day, month, and year from the Ethiopian calendar to Gregorian. It allows researchers working with a large data set to easily convert date.
}
\author{
\strong{Maintainer}: Dawit Jember \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
