% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convdate.R
\name{davedate}
\alias{davedate}
\title{Convert date from the Ethiopian calendar to the Gregorian}
\usage{
davedate(a, b, c, id)
}
\arguments{
\item{a}{a numeric value of day DD}

\item{b}{a numeric value of month MM}

\item{c}{a numeric value of year YYYY}

\item{id}{a numeric value of individual}
}
\value{
a date value
}
\description{
Convert date from the Ethiopian calendar to the Gregorian
}
\examples{
#To convert date, month, year from a data set

davedate(a=data$day, b=data$month, c=data$year, id=data$id)
}
